Adapt community roles to current practices
Some information in the file weren't accurate (e.g. the
`@fluxcd/member` GitHub team doesnt exist) so I did my best in
adapting the file to current practice.

Please take this as more of a base for discussing these and probably
further changes to the role definitions and processes mentioned in the
file.

Signed-off-by: Max Jonas Werner <mail@makk.es>