Step down as maintainer for Flux Community
Unfortunately it's already getting clear that I won't have time to
continue serving in this capacity, so I'll demote myself from
maintainer to project member.

It has been a big pleasure and honour to work alongside all of you
and I'm very proud of what we achieved together.

Please don't hesitate to get in touch, if I can be of any help.
I'll check in on the Slacks every now and then...

Signed-off-by: Daniel Holbach <daniel.holbach@gmail.com>