Add KUBECON.md
This is associated with fluxcd/website#1804

It will not be published on the website before we merge
fluxcd/website#1804

We can see that it works in the preview build, (but we will need to
merge this one first before we can see it working in the preview.)

I don't see why we would publish this page any earlier than ~1-2 weeks
before KubeCon at earliest.

But there may be no reason to hold it up, either!

Signed-off-by: Kingdon P Barrett <kingdon@tuesdaystudios.com>