Not bad (Update KUBECON.md)
Rearrange clearfix divs

This is now looking like an actual serviceable design

It replicates the original layout. It is a little bit brittle, but it's
not near as horrible as I thought it would be. We can add CSS as needed,
(the PR fluxcd/website#1804 will get the needed CSS updates to match
these classes)

sed ggEz

Signed-off-by: Kingdon Barrett <kingdon@tuesdaystudios.com>