Push to Website workflow
Upon any change in the main branch of any of the files from community
that we know to be listed in external-sources/ on fluxcd/website, we
should trigger the Netlify job so that change is rebuilt immediately.

Without this, there is the problem that no reasonably long polling
interval is going to satisfy UX and DX expectations, because we cannot
subscribe to updates from Google Sites.

(We must poll for changes, upon detecting the changes scrape them into a
PR, wait for someone to merge the PR, then trigger the next workflow. If
we set an interval of 6h then the maximum wait time to see your change
reflected is 12h, but really it's much longer because there is some lag
time while you wait for community maintainer to notice that PR, approve
and merge it!)

The problem can also be addressed by shortening the polling interval,
but that only raises the perceived impact of waiting for a maintainer to
notice the PR, approve, and merge it. We could shorten both polling
intervals, and that would be very inefficient by comparison. And we
would still always have to wait at least as long as it takes for the
maintainer to come back after the polling interval.

Instead, we can just shorten the update-kubecon workflow interval, if
necessary, for the duration of the conference - then trigger the remote
workflow and there is no need to shorten both intervals. Merging this PR
will create a quality-of-life improvement for the community and website
maintainers.

NB: someone should set the update interval back to every 6 hours after
the conference crunch is over.

Signed-off-by: Kingdon P Barrett <kingdon@tuesdaystudios.com>